Increased precision

+++ ufomap_ros/ufomap_rviz_plugins/include/ufomap_rviz_plugins/color.h
@@ -57,8 +57,8 @@ static inline constexpr auto color_lut = [] {
 	std::array<float, std::numeric_limits<ufo::map::RGBColorType>::max() + 1> color_lut =
 	    {};
 	for (size_t i = 0; i != color_lut.size(); ++i) {
-		color_lut[i] = static_cast<float>(i) /
-		               static_cast<float>(std::numeric_limits<ufo::map::RGBColorType>::max());
+		color_lut[i] = static_cast<double>(i) /
+		               static_cast<double>(std::numeric_limits<ufo::map::RGBColorType>::max());
 	}
 	return color_lut;
 }();
